+++ CHANGELOG.md
@@ -10,10 +10,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-## [0.5.3]
+## [0.6.0]
+
+### Added
+- Support for Anthropic's text_delta format in content extraction in case thinking is enabled (set like this `cb = StreamCallback(; out = stdout, flavor = AnthropicStream(), kwargs = (; include_thinking = false))` to see it). We now let all index chunks pass through.
 
 ### Fixed
-- Support for Anthropic's text_delta format in content extraction in case thinking is enabled
+- Fixed assertion for content-type to be more informative when you hit limits or some model error (see the status code and response headers).
 
 ## [0.5.2]